Link blogs to users and implement token-based authentication

+++ controllers/blogs.js
@@ -1,37 +1,62 @@
+const jwt = require('jsonwebtoken')
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const middleware = require('../utils/middleware')
 
 blogsRouter.get('/', async (request, response, next) => {
     try {
-        const blogs = await Blog.find({})
+        const blogs = await Blog.find({}).populate('user', { username: 1, name: 1})
         response.status(200).json(blogs)
     } catch (error) {
         next(error)
     }
 })
 
-blogsRouter.post('/', async (request, response, next) => {
-    const blog = new Blog(request.body)
+blogsRouter.post('/', middleware.userExtractor, async (request, response, next) => {
+    const { title, author, url, likes } = request.body
+    const user = request.user
 
     try {
+        const blog = new Blog({
+            title: title,
+            author: author,
+            url: url,
+            likes: likes,
+            user: user.id
+        })
+
         const savedBlog = await blog.save()
+        user.blogs = user.blogs.concat(savedBlog._id)
+        await user.save()
         response.status(201).json(savedBlog)
     } catch (error) {
         next(error)
     }
 })
 
-blogsRouter.delete('/:id', async(request, response, next) => {
+blogsRouter.delete('/:id', middleware.userExtractor, async(request, response, next) => {
+    if (!request.token) {
+        return response.status(401).json({ error: 'no token' })
+    }
+
     const id = request.params.id
+    const user = request.user
 
     try {
-        const result = await Blog.findByIdAndDelete(id)
+        const blog =  await Blog.findById(id)
 
-        if (!result) {
-            response.status(404).json({ error: 'Blog not found' });
-        } else {
-            response.status(204).end();
+        if (!blog) {
+            response.status(404).json({ error: 'Blog not found' })
         }
+
+        if (user.id.toString() !== blog.user.toString()) {
+            response.status(400).json({ error: 'invalid user' })
+        }
+
+        await Blog.findByIdAndDelete(id)
+
+        response.status(204).end()
     } catch (error) {
         next(error)
     }
